Refactor stack_test.go and fix error in Test_Stack()
The changes made in this commit involve refactoring the stack_test.go file and fixing an error in the Test_Stack() function. The error was causing the test to fail when the expected and actual results did not match. The commit message includes a concise and informative description of the changes made.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>