feat(linter): Add more linter to the codebase (#147)
* chore(golangci): add more linters

This commit adds an exclusion for the "Added to album" issues in the `.golangci.yml` file. This change allows the linter to ignore these specific issues during the code analysis.

The commit also includes additional linters such as `stylecheck`, `unused`, `goconst`, `makezero`, `unparam`, `prealloc`, `predeclared`, `unconvert`, and `whitespace` to enhance the code quality analysis.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* Refactor stack_test.go and fix error in Test_Stack()

The changes made in this commit involve refactoring the stack_test.go file and fixing an error in the Test_Stack() function. The error was causing the test to fail when the expected and actual results did not match. The commit message includes a concise and informative description of the changes made.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

* Refactor condition checks for improved readability and efficiency

This commit introduces changes to the codebase in the `immich/call.go` and `main.go` files. The modifications involve utilizing a shorter and more readable syntax for condition checks in one file, and a more efficient and concise syntax for condition checks in the other file. Specifically, the changes involve replacing the length check of a string variable with a check for non-empty string.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>

---------

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>