update create time when remote does not have one
I uploaded my takeout directly naively, so all the times were incorrect. I updated this tool to be able to detect when the local and remote file are the same, but the local timestamp is *older* then the remote one, which indicates that the local one is better because bad timestamps seem to always be newer than real ones as they are fabricated from when downloaded from google or when uploaded to immich

There were other bugs that needed fixing for this--OriginalFileName already had the extension, so I removed the + ext. Also, the updateAsset API seems to have been updated since you added it.